+++ test/main-spec.js
@@ -125,6 +125,13 @@ describe('promised-handlebars:', function () {
       .to.eventually.equal('h(aa),h(aa)')
       .notify(done)
   })
+  it('async helpers nested in synchronous builtin block-helpers should work', function (done) {
+    var template = Handlebars.compile(fixture('builtin-block-helper-nests-async.hbs'))
+    return expect(template({ arr: [{ a: 'aa'}, { a: 'bb'}]}))
+      .to.eventually.equal('h(aa)-h(bb)-')
+      .notify(done)
+  })
+
 })
 
 function fixture (file) {