+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Upcoming
+### Fix 
+* Make sure that builtin block-helpers are also wrapped appropriately to work with
+  promise-helpers inside their blocks
+
 ## v0.3.2 - 2015-07-30
 ### Fix